third_party/blink/web_tests/wpt_internal/attribution-reporting/simple-aggregatable-report-test.sub.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
promise_test(async t => {
  const interval = 100;
  const source = {
    source_event_id: '9153495207648170476',
    destination: `https://{{host}}`,
    aggregation_keys: {
      campaignCounts: "0x159",
      geoValue: "0x5",
    }
  };
  const trigger = {
    aggregatable_trigger_data: [
      {
        key_piece: '0x400',
        source_keys: ['campaignCounts'],
      },
      {
        key_piece: '0xA80',
        source_keys: ['geoValue', 'nonMatchingKey'],
      }
    ],
    aggregatable_values: {
      campaignCounts: 32768,
      geoValue: 1664,
    },
  };
  await resetAggregatableReports();
  registerAttributionSrc('Attribution-Reporting-Register-Source', source);
  // Adds delay to reduce chance of race condition between source registration and
  await delay(interval);
  // trigger registration.
  registerAttributionSrc('Attribution-Reporting-Register-Trigger', trigger);
  const payload = await pollAggregatableReports(interval);
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0]);
  assert_own_property(report, 'shared_info');
  const shared_info = JSON.parse(report.shared_info);
  assert_own_property(shared_info, 'api');
  assert_equals(shared_info.api, 'attribution-reporting');
  assert_own_property(shared_info, 'report_id');
  assert_own_property(shared_info, 'reporting_origin');
  assert_own_property(shared_info, 'scheduled_report_time');
  assert_own_property(shared_info, 'version');
  assert_own_property(shared_info, 'attribution_destination');
  assert_equals(shared_info.attribution_destination, source.destination);
  assert_not_own_property(report, 'source_debug_key');
  assert_not_own_property(report, 'trigger_debug_key');
  assert_own_property(report, 'aggregation_service_payloads');
  assert_equals(report.aggregation_service_payloads.length, 1);
  const aggregation_service_payload = report.aggregation_service_payloads[0];
  assert_own_property(aggregation_service_payload, 'payload');
  assert_own_property(aggregation_service_payload, 'key_id');
  assert_not_own_property(aggregation_service_payload, 'debug_cleartext_payload');
}, 'Ensure aggregatable attribution report is received.');
</script>
